Switch to store for cart so user can change page and loose the current cart
