+++ pages/admin/sales/new.vue
@@ -3,11 +3,11 @@
   import SalePaymentModeQuery from "~/composables/api/query/SalePaymentModeQuery";
   import type {InventoryItem} from "~/types/inventoryItem";
   import {formatMonetary} from "~/utils/string";
-  import type {FormError} from "#ui/types";
   import type {SalePaymentMode} from "~/types/salePaymentMode";
   import SaleQuery from "~/composables/api/query/SaleQuery";
   import type {Sale} from "~/types/sale";
   import type {SalePurchasedItem} from "~/types/salePurchasedItem";
+  import {useCartStore} from "~/stores/useCartStore";
 
   definePageMeta({
     layout: "pos"
@@ -22,12 +22,14 @@
 
   const toast = useToast()
 
+  const cartStore = useCartStore()
+  const { searchQuery, cart, cartTotalPrice, cartComment, cartCustomItemModalOpen, customItemForm, selectedPaymentMode } = storeToRefs(cartStore)
+
   const inventoryItemQuery = new InventoryItemQuery()
   const paymentModeQuery = new SalePaymentModeQuery()
   const saleQuery = new SaleQuery()
 
-  const searchQuery: Ref<string> = ref('')
-  const searchQueryInput: Ref<string> = ref('')
+  const searchQueryInput: Ref<string> = ref(searchQuery.value)
   watch(searchQuery, () => {
     searchQueryInput.value = searchQuery.value
     loadItems()
@@ -57,24 +59,8 @@
     return categories
   })
 
-  const cart: Ref<Map<string, {item: InventoryItem, quantity: number}>> = ref(new Map())
-  const cartTotalPrice = computed( () => {
-    let total: number = 0
-    cart.value.forEach( item => {
-      total += item.quantity * Number(item.item.sellingPrice)
-    } )
-    return !isNaN(total) ? total.toFixed(2) : '0.00'
-  } )
-  const cartComment = ref('')
-  const customItemForm = reactive<{name: string|undefined, sellingPrice: string|undefined}>({
-    name: undefined,
-    sellingPrice: undefined
-  })
-  const cartCustomItemModalOpen = ref(false)
-
   // Payment mode
   const paymentModes: Ref<SalePaymentMode[]> = ref([])
-  const selectedPaymentMode: Ref<SalePaymentMode | undefined> = ref(undefined)
 
   async function loadItems(page: number = 1) {
     isLoading.value = true
@@ -126,68 +112,6 @@
   const cameraPreview = ref(false)
   const cameraIsPresent = verifyCameraIsPresent()
 
-  // Cart management
-
-  const validateCustomItemForm = (state: any): FormError[] => {
-    const errors = []
-    if (!state.name) errors.push({ path: 'name', message: 'Champ requis' })
-    if (!state.sellingPrice) errors.push({ path: 'sellingPrice', message: 'Champ requis' })
-    const sellingPrice = parseFloat(state.sellingPrice)
-    if (isNaN(sellingPrice)) errors.push({ path: 'sellingPrice', message: 'Le champ doit être un nombre' })
-    return errors
-  }
-
-  function addToCart(item: InventoryItem, modifier: number = 1) {
-    if (!item.id) {
-      return
-    }
-
-    searchQuery.value = '' // Since the item has been added to the cart, we revert to empty search
-
-    let cartItem = cart.value.get(item.id.toString())
-
-    if (!cartItem) {
-      cartItem = { quantity: modifier, item: item }
-    } else {
-      cartItem.quantity += modifier
-    }
-
-    if (cartItem.quantity <= 0) {
-      cart.value.delete(item.id.toString())
-    } else {
-      cart.value.set(item.id.toString(), cartItem)
-    }
-  }
-
-  function emptyCart() {
-    cart.value.clear()
-    searchQuery.value = ''
-    selectedPaymentMode.value = undefined
-    cartComment.value = ''
-  }
-
-  function addCustomItemToCart() {
-    if (!customItemForm.sellingPrice) {
-      closeCustomItemModal()
-      return
-    }
-
-    // Negative ID, Backend know it's not a real item
-    const item: InventoryItem = {
-      id: - Math.floor(Math.random() * 200000),
-      name: customItemForm.name,
-      sellingPrice: parseFloat(customItemForm.sellingPrice.replace(',', '.')).toFixed(2),
-    }
-    addToCart(item)
-    closeCustomItemModal()
-  }
-
-  function closeCustomItemModal() {
-    customItemForm.name = undefined
-    customItemForm.sellingPrice = undefined
-    cartCustomItemModalOpen.value = false
-  }
-
   // Sale management
   async function createSale() {
     isCreatingSale.value = true
@@ -222,7 +146,7 @@
       toast.add({
         color: "red",
         title: "La vente à échoué",
-        description: error.message
+        description: error?.message
       });
       return;
     }
@@ -231,6 +155,8 @@
       color: "green",
       title: "Vente enregistrée",
     });
+
+    cartStore.emptyCart()
 
     navigateTo('/admin/sales/' + created.id)
   }
@@ -294,7 +220,7 @@
               <div v-if="item.description" class="text-xs">{{ item.description }}</div>
             </div>
             <div class="text-xs bg-neutral-200 dark:bg-gray-800 p-1 rounded-md">{{ formatMonetary(item.sellingPrice) }}</div>
-            <UButton icon="i-heroicons-shopping-cart" size="2xs" @click="addToCart(item)" />
+            <UButton icon="i-heroicons-shopping-cart" size="2xs" @click="cartStore.addToCart(item)" />
           </div>
         </div>
 
@@ -311,7 +237,7 @@
 
       <UModal v-model="cartCustomItemModalOpen">
         <UCard>
-          <UForm class="flex gap-2 flex-col" :state="customItemForm" :validate="validateCustomItemForm" @submit="addCustomItemToCart">
+          <UForm class="flex gap-2 flex-col" :state="customItemForm" :validate="cartStore.validateCustomCartForm" @submit="cartStore.addCustomItemToCart">
             <UFormGroup label="Nom" name="name">
               <UInput v-model="customItemForm.name"/>
             </UFormGroup>
@@ -325,7 +251,7 @@
             </UFormGroup>
 
             <div class="flex gap-2 mt-2 justify-end">
-              <UButton color="red" variant="ghost" @click="closeCustomItemModal()">
+              <UButton color="red" variant="ghost" @click="cartStore.closeCustomItemModal()">
                 Annuler
               </UButton>
 
@@ -344,13 +270,13 @@
         <div class="mt-4">
           <div class="flex text-xs align-center mt-1">
             <div class="flex-1"></div>
-            <UButton v-if="cart.size > 0" size="xs" @click="emptyCart()">Vider le panier</UButton>
+            <UButton v-if="cart.size > 0" size="xs" @click="cartStore.emptyCart()">Vider le panier</UButton>
           </div>
           <div v-if="cart.size < 1" class="text-center">
             <i>Aucun articles</i>
           </div>
           <div v-for="[id , cartItem] in cart" class="flex items-center gap-2 mb-1">
-            <GenericStackedUpDown @changed="modifier => { addToCart(cartItem.item, modifier) }" />
+            <GenericStackedUpDown @changed="modifier => { cartStore.addToCart(cartItem.item, modifier) }" />
 
             <div class="text-xs bg-neutral-200 dark:bg-gray-800 p-1 rounded-md">{{ cartItem.quantity }}</div>
             <div class="text-sm flex-1 leading-tight">{{ cartItem.item.name }}</div>
